LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/subtree-exclusion-expected.txt


FAIL Subtree exclusion with overflow-anchor. assert_equals: expected 200 but got 150
